backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
distlib==0.3.7
Django==4.2.6
django-debug-toolbar==4.2.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
filelock==3.12.4
gunicorn==21.2.0
identify==2.5.30
idna==3.4
inflection==0.5.1
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
platformdirs==3.11.0
pre-commit==3.4.0
psycopg2-binary==2.9.8
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.24.5
